SessionHelperにcurrent_userを切り出した。
