+++ app/apis/api/v1/posts.rb
@@ -5,24 +5,8 @@ module API
       # https://stackoverflow.com/questions/35344117/why-does-session-doesnt-work-in-grape-rails
       use ActionDispatch::Session::CookieStore, key: 'user_id'
 
+      helpers API::V1::Helpers::SessionHelper
       helpers do
-        def session
-          env['rack.session']
-        end
-
-        # TODO: sessionに保持しておくものがuse_idでは実用的ではないので
-        # あとで必ず書き換える必要がある。
-        # faowighoiwh98haf0anwifahwioa
-        # こんな感じのトークンになるように
-        def current_user
-          # TODO: あとで必ず修正する
-          # sessionsのAPIを作るまではcurrent_uesrを参照したら現在のユーザー1のuser_idを設定するようにする
-          session[:user_id] = 1 if Rails.env.development?
-          return unless session[:user_id]
-          return @current_user if defined?(@current_user)
-          @current_user = ::User.find(session[:user_id])
-        end
-
         def post_with_revision_params
           ActionController::Parameters.new(customize_params).permit(
             revisions_attributes: %i[
